ADD: minimum detectable signal
Adding the Fortran routine that objectively determines the noise floor
of a power signal. For example, the input field can be reflectivity or
signal-to-noise ratio. This routine is based on the method of Hildebrand
and Sekhon (1974). As of now the routine expects the power signal to be
sorted in descending order for each range.