Testing new mapping routine
A couple nights ago I was playing around with scipy.spatial.cKDTree and
discovered that it was extremely fast (~ seconds) in looking up nearest
neighbors when (1) the analysis grid was large (10^6) and (2) the data was
large (10^5) and (3) the number of nearest neighbors to search for was
large (~ 100). Therefore I'm thinking that the current implementation of
the mapping routine could be greatly improved in terms of processing time.
I think where the current implementation fails is in looping over the
analysis grid, i.e. looping over 10^6 analysis points. I believe we do
this because often times a variable radius of influence is required when
mapping (interpolating) scanning radar data, and therefore each grid point
(potentially) requires a new search radius for the k-d tree. However in my
testing I tried a value of 50 for the number of nearest neighbors to
search for, and it took just a few seconds to run. Since one output of
SciPy's k-d tree query is distance, then I think we should be tackling the
problem slightly different by first looking for a specified number of nearest
neighbors, and then removing (masking) those that are outside a specified
radius. More to come...