Implement Cressman scheme
I am committing some additions that (hopefully) implement the Cressman
scheme. The key here is how to handle neighbors that are not within the
radius of influence (ROI) and neighbors that are not within the
distance_upper_bound parameter. These two criteria are controlled by the
number of nearest neighbors (k) to return. In general, k >> 1 (e.g., 100),
however the larger k is, the longer it takes to search the k-d tree.
However, right now I am under the impression this increase in search time
is rather insignificant when compared to how long the old mapping routine
took. I believe the most logical way to handle the above two criteria is
to find where the distance from a neighbor is larger than the ROI (which
implies larger than distance_upper_bound) and set the weight for this
neighbor to zero so that it does not affect the interpolation (weighted
average) step. More to come...