Beginning to implement further gate removal
I am beginning to implement further gate removal when Cressman and
Barnes weights are desired. This is based on the fact that for a (1)
Barnes scheme there is generally a "cutoff radius" and (2) Cressman
scheme there is a maximum radius of influence. These two facts allow us
to remove radar gates that will never be captured by an analysis grid
point. More to come...