Putting on the final touches...
I made some changes to the output grid object, as well as renamed the
main function map_to_grid -> map_radar_to_grid so that it is more
specific (i.e. as of now only one radar can be mapped at a time). The
axes of the output grid object are now compatible with write_grid.