MOD: minimum signal detection routine
Adding another subroutine to my minimum detection Fortran module. This
was motivated by the fact that the current subroutine, though producing
reasonable results, is not capable of identifying the close range gates
for lower elevations as noise, which they undoubtedly are just from
inspection. Therefore I created another subroutine which goes sweep by
sweep and determines the minimum detectable signal (MDS). I ran some
experiments with this and the MDS results for some sweeps are not
physically consistent, i.e., MDS does not strictly increase as a
function of range. More needs to be done...